feat(network): refine network flow and allow selecting remote address (#22)
Major rework. Decouples Backend and Client/Game.
Now, game clients automatically connect to their own backend and an additional menu button allows one to also join another server. In the future, joining a server that is not their own backend shall lead to the backend connecting to the server instead, relaying and synchronising data and bundling this data for the game client