feat(rpc): give server getter and setter for spaceship data (#31)
NOTE: This PR gives the server control over spaceships. It will allow the server to request current pos/vel/angle and to set it.
Getting data is necessary to sync a spaceship up with the remote one of a newly joined client.
Setting data will soon be necessary to allow the server to sync data in more or less regular intervals.