feat(p2p): cross-share connections on join (#44)
NOTES: Additionally, adds an in-game feat to join other clients while the game is running. This is only for testing, and should be removed later on. This feature has no way of verifying the rng seed, and I do not plan to implement one. Hence, join must not be used with differing seeds